+++ analysis_options.yaml
@@ -1,10 +1,44 @@
 include: package:lints/recommended.yaml
 
-linter:
-  rules:
-    - avoid_dynamic_calls
-
 analyzer:
   exclude:
     - samples/large/**
     - spec/**
+  language:
+    # strict-casts: true
+    # strict-inference: true
+  errors:
+    constant_identifier_names: ignore
+    non_constant_identifier_names: ignore
+
+linter:
+  rules:
+    # consistency
+    - combinators_ordering
+    - directives_ordering
+    - library_annotations
+    # - lines_longer_than_80_chars
+    # - omit_local_variable_types
+    - prefer_asserts_in_initializer_lists
+    # - prefer_const_constructors
+    - prefer_relative_imports
+    - prefer_single_quotes
+    - sort_pub_dependencies
+    # - unnecessary_lambdas
+    - unnecessary_library_directive
+    - unnecessary_parenthesis
+    - unnecessary_statements
+    - use_is_even_rather_than_modulo
+
+    # correctness
+    - always_declare_return_types
+    # - avoid_catching_errors
+    - avoid_dynamic_calls
+    - comment_references
+    - conditional_uri_does_not_exist
+    # todo: add this back
+    # - only_throw_errors
+    - test_types_in_equals
+    - throw_in_finally
+    - type_annotate_public_apis
+    - unawaited_futures
