+++ lib/module.dart
@@ -1,0 +1,3 @@
+// import 'loader.dart';
+
+export 'loader.dart' show Module;
