+++ samples/print_main.dart
@@ -1,4 +1,4 @@
-import 'package:wasmi/runtime.dart';
+import 'package:wasmi/old/runtime.dart';
 
 import 'print.dart';
