--- test/src/spectest.wat
@@ -1,18 +1,0 @@
-(module
-  (global (export "global_i32") i32 (i32.const 666))
-  (global (export "global_i64") i64 (i64.const 666))
-  (global (export "global_f32") f32 (f32.const 0))
-  (global (export "global_f64") f64 (f64.const 0))
-
-  (table (export "table") 10 20 funcref)
-
-  (memory (export "memory") 1 2)
-
-  (func (export "print"))
-  (func (export "print_i32") (param i32))
-  (func (export "print_i64") (param i64))
-  (func (export "print_f32") (param f32))
-  (func (export "print_f64") (param f64))
-  (func (export "print_i32_f32") (param i32 f32))
-  (func (export "print_f64_f64") (param f64 f64))
-)
