--- test/unit/blocks.wat
@@ -1,37 +1,0 @@
-(module
-  (func (export "nested-block-value-1") (param i32) (result i32)
-    (i32.add
-      (i32.const 1)
-      (block (result i32)
-        (drop (i32.const 2))
-        (i32.add
-          (i32.const 4)
-          (block (result i32)
-            (drop (br_if 1 (i32.const 8) (local.get 0)))
-            (i32.const 16)
-          )
-        )
-      )
-    )
-  )
-
-  (func (export "nested-block-value-2") (result i32)
-    (i32.add
-      (i32.const 1)
-      (block (result i32)
-        (call $dummy)
-        (i32.add (i32.const 4) (br 0 (i32.const 8)))
-      )
-    )
-  )
-
-  (func (export "params") (result i32)
-    (i32.const 1)
-    (i32.const 2)
-    (block (param i32 i32) (result i32)
-      (i32.add)
-    )
-  )
-
-  (func $dummy)
-)
