--- test/unit/blocks_test.dart
@@ -1,17 +1,0 @@
-import 'package:test/test.dart';
-
-import 'blocks.dart';
-
-void main() {
-  group('blocks', () {
-    late BlocksModule module;
-
-    setUp(() {
-      module = BlocksModule();
-    });
-
-    test('call params()', () {
-      expect(module.params(), 3);
-    });
-  });
-}
