--- tool/tests.md
@@ -1,92 +1,0 @@
-| Spec test                   | Generation status |
-| ---                         | --- |
-| address.wast                | +   |
-| align.wast                  | +   |
-| block.wast                  | -   |
-| br_if.wast                  | -   |
-| br_table.wast               | -   |
-| br.wast                     | -   |
-| call_indirect.wast          | -   |
-| call.wast                   | -   |
-| comments.wast               | +   |
-| const.wast                  | +   |
-| conversions.wast            | +   |
-| endianness.wast             | +   |
-| exports.wast                | +   |
-| f32_bitwise.wast            | +   |
-| f32_cmp.wast                | +   |
-| f32.wast                    | +   |
-| f64_bitwise.wast            | +   |
-| f64_cmp.wast                | +   |
-| f64.wast                    | +   |
-| fac.wast                    | -   |
-| float_exprs.wast            | +   |
-| float_literals.wast         | +   |
-| float_memory.wast           | +   |
-| float_misc.wast             | +   |
-| forward.wast                | -   |
-| func_ptrs.wast              | +   |
-| func.wast                   | -   |
-| global.wast                 | +   |
-| i32.wast                    | +   |
-| i64.wast                    | +   |
-| if.wast                     | -   |
-| int_exprs.wast              | +   |
-| int_literals.wast           | +   |
-| labels.wast                 | -   |
-| left-to-right.wast          | +   |
-| load.wast                   | -   |
-| local_get.wast              | -   |
-| local_set.wast              | +   |
-| local_tee.wast              | -   |
-| loop.wast                   | -   |
-| memory_copy.wast            | +   |
-| memory_fill.wast            | +   |
-| memory_grow.wast            | -   |
-| memory_init.wast            | +   |
-| memory_redundancy.wast      | +   |
-| memory_size.wast            | +   |
-| memory_trap.wast            | +   |
-| memory.wast                 | +   |
-| nop.wast                    | -   |
-| ref_func.wast               | +   |
-| ref_is_null.wast            | +   |
-| ref_null.wast               | +   |
-| return.wast                 | -   |
-| select.wast                 | -   |
-| stack.wast                  | -   |
-| start.wast                  | -   |
-| store.wast                  | -   |
-| switch.wast                 | -   |
-| table_copy.wast             | +   |
-| table_fill.wast             | +   |
-| table_get.wast              | +   |
-| table_grow.wast             | +   |
-| table_init.wast             | -   |
-| table_set.wast              | +   |
-| table_size.wast             | +   |
-| traps.wast                  | -   |
-| type.wast                   | -   |
-| unreachable.wast            | -   |
-| unreached-valid.wast        | -   |
-| unwind.wast                 | -   |
-| bulk.wast                   | properly support FuncRef instruction initializers |
-| data.wast                   | need to rework DataSegments initializers |
-| elem.wast                   | requires reworking ElementSegments initialization |
-| imports.wast                | duplicate symbol imports |
-| linking.wast                | duplicate symbol imports |
-| names.wast                  | not completely working |
-| binary-leb128.wast          | (no tests) |
-| binary.wast                 | (no tests) |
-| custom.wast                 | (no tests) |
-| inline-module.wast          | (no tests) |
-| skip-stack-guard-page.wast  | (no tests) |
-| table-sub.wast              | (no tests) |
-| table.wast                  | + (no tests) |
-| token.wast                  | (no tests) |
-| tokens.wast                 | (no tests) |
-| unreached-invalid.wast      | (no tests) |
-| utf8-custom-section-id.wast | (no tests) |
-| utf8-import-field.wast      | (no tests) |
-| utf8-import-module.wast     | (no tests) |
-| utf8-invalid-encoding.wast  | (no tests) |
